refactor: Cleaner configuration of empty state in fare contract list
This refactor started when I wanted to make the 'mode' parameter
non-optional, since it wasn't intutive what the default values was.
Making it required made it obvious that the possible modes (historic
and sent) was not enough as in the my tickets tab the neither
historic or sent was fitting.

All in all I decided to remove the mode parameter and make all of the
empty state config parametrized from the outside. This also removed
the circular dependency where ticket history screen used fare contract
list, and the fare contract list used the TicketHistoryMode type from
the ticket history screen.
