+++ src/fare-contracts/FareContractAndReservationsList.tsx
@@ -4,13 +4,12 @@ import {FareContractOrReservation} from '@atb/fare-contracts/FareContractOrReser
 import {FareContract, Reservation, TravelCard} from '@atb/ticketing';
 import {NavigationProp, useNavigation} from '@react-navigation/native';
 import {useAnalyticsContext} from '@atb/analytics';
-import {HoldingHands, TicketTilted} from '@atb/assets/svg/color/images';
 import {EmptyState} from '@atb/components/empty-state';
-import {TicketHistoryMode} from '@atb/ticket-history';
 import {useSortFcOrReservationByValidityAndCreation} from './utils';
 import {getFareContractInfo} from './utils';
 import {StyleSheet} from '@atb/theme';
 import {View} from 'react-native';
+import type {EmptyStateProps} from '@atb/components/empty-state';
 
 type RootNavigationProp = NavigationProp<RootStackParamList>;
 
@@ -19,18 +18,17 @@ type Props = {
   fareContracts?: FareContract[];
   now: number;
   travelCard?: TravelCard;
-  mode?: TicketHistoryMode;
-  emptyStateTitleText: string;
-  emptyStateDetailsText: string;
+  emptyStateConfig: Pick<
+    EmptyStateProps,
+    'title' | 'details' | 'illustrationComponent'
+  >;
 };
 
 export const FareContractAndReservationsList: React.FC<Props> = ({
   fareContracts,
   reservations,
   now,
-  mode = 'historical',
-  emptyStateTitleText,
-  emptyStateDetailsText,
+  emptyStateConfig,
 }) => {
   const styles = useStyles();
   const navigation = useNavigation<RootNavigationProp>();
@@ -50,12 +48,7 @@ export const FareContractAndReservationsList: React.FC<Props> = ({
   return (
     <View style={styles.container}>
       {!fareContractsAndReservationsSorted.length && (
-        <EmptyState
-          title={emptyStateTitleText}
-          details={emptyStateDetailsText}
-          illustrationComponent={emptyStateImage(mode)}
-          testID="fareContracts"
-        />
+        <EmptyState {...emptyStateConfig} testID="fareContracts" />
       )}
       {fareContractsAndReservationsSorted?.map((fcOrReservation, index) => (
         <FareContractOrReservation
@@ -76,15 +69,6 @@ export const FareContractAndReservationsList: React.FC<Props> = ({
       ))}
     </View>
   );
-};
-
-const emptyStateImage = (emptyStateMode: TicketHistoryMode) => {
-  switch (emptyStateMode) {
-    case 'historical':
-      return <TicketTilted height={84} />;
-    case 'sent':
-      return <HoldingHands height={84} />;
-  }
 };
 
 const useStyles = StyleSheet.createThemeHook((theme) => ({
