+++ src/stacks-hierarchy/Root_TabNavigatorStack/TabNav_TicketingStack/Ticketing_TicketTabNavStack/TicketTabNav_AvailableFareContractsTabScreen.tsx
@@ -14,6 +14,7 @@ import {usePopOverContext} from '@atb/popover';
 import {useOneTimePopover} from '@atb/popover/use-one-time-popover';
 import {isElementFullyInsideScreen} from '@atb/utils/is-element-fully-inside-screen';
 import {TravelTokenBox} from '@atb/travel-token-box';
+import {TicketTilted} from '@atb/assets/svg/color/images';
 
 type Props =
   TicketTabNavScreenProps<'TicketTabNav_AvailableFareContractsTabScreen'>;
@@ -105,14 +106,17 @@ export const TicketTabNav_AvailableFareContractsTabScreen = ({
           reservations={reservations}
           fareContracts={availableFareContracts}
           now={serverNow}
-          emptyStateTitleText={t(
-            TicketingTexts.availableFareProductsAndReservationsTab
-              .noActiveTicketsTitle,
-          )}
-          emptyStateDetailsText={t(
-            TicketingTexts.availableFareProductsAndReservationsTab
-              .noActiveTicketsDetails,
-          )}
+          emptyStateConfig={{
+            title: t(
+              TicketingTexts.availableFareProductsAndReservationsTab
+                .noActiveTicketsTitle,
+            ),
+            details: t(
+              TicketingTexts.availableFareProductsAndReservationsTab
+                .noActiveTicketsDetails,
+            ),
+            illustrationComponent: <TicketTilted height={84} />,
+          }}
         />
         <Section>
           {hasHistoricalFareContracts && (
